# Improve treeHasher performance