@cuonglm 2021-07-13
> LGTM, thank you @cuonglm! Let's also see if we can shave off by reusing the hashers

Yeah, I think it can, but want to keep the PR as small as possible.